extract all functions as helpers
