+++ src/assets/theme/helpers/color-utils.ts
@@ -14,6 +14,9 @@
  * limitations under the License.
  */
 
+// chroma-js is a library for all kinds of color conversions and color scales.
+import chroma, { Color } from 'chroma-js';
+
 import { GreyColorName, PaletteColorKey, PaletteOptions } from '@mui/material';
 
 export const paletteKeys: (keyof PaletteOptions)[] = [
@@ -52,3 +55,44 @@ export const isGreyColorName = (color: string): color is GreyColorName =>
   color === 'grey-700' ||
   color === 'grey-800' ||
   color === 'grey-900';
+
+/**
+ =========================================================
+ * Material Kit 2 React - v2.0.0
+ =========================================================
+
+ * Product Page: https://www.creative-tim.com/product/material-kit-react
+ * Copyright 2021 Creative Tim (https://www.creative-tim.com)
+
+ Coded by www.creative-tim.com
+
+ =========================================================
+
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ */
+
+/**
+ The hexToRgb() function helps you to change the hex color code to rgb
+ using chroma-js library.
+ */
+
+export const hexToRgb = (color: string | number | Color): string =>
+  chroma(color).rgb(false).join(', ');
+
+/**
+ The rgba() function helps you to create a rgba color code, it uses the hexToRgb() function
+ to convert the hex code into rgb for using it inside the rgba color format.
+ */
+
+export const rgba = (color: string | number | Color, opacity: number): string =>
+  `rgba(${hexToRgb(color)}, ${opacity})`;
+
+/**
+ The linearGradient() function helps you to create a linear gradient color background
+ */
+
+export const linearGradient = (
+  color: string | number | Color,
+  colorState: string | number | Color,
+  angle = 195,
+): string => `linear-gradient(${angle}deg, ${color}, ${colorState})`;
