+++ src/types/@mui-material-styles.d.ts
@@ -20,7 +20,6 @@ import {
   TypographyUtils,
   Variant,
 } from '@mui/material/styles/createTypography';
-import { Color } from 'chroma-js';
 
 declare module '@mui/material/styles' {
   interface BreakpointOverrides {
@@ -81,27 +80,8 @@ declare module '@mui/material/styles' {
     inset: string;
   };
 
-  type Functions = {
-    boxShadow: (
-      offset?: number[],
-      radius?: number[],
-      color: string | number | Color,
-      opacity: number,
-      inset?: string,
-    ) => string;
-    hexToRgb: (color: string | number | Color) => string;
-    linearGradient: (
-      color: string | number | Color,
-      colorState: string | number | Color,
-      angle?: number,
-    ) => string;
-    pxToRem: (number: number, baseNumber?: number) => string;
-    rgba: (color: string | number | Color, opacity: number) => string;
-  };
-
   interface Theme {
     palette: Palette & SocialPalette;
-    functions: Functions;
   }
 
   interface Palette {
@@ -131,7 +111,6 @@ declare module '@mui/material/styles' {
   // allow configuration using `createTheme`
   interface ThemeOptions {
     palette?: PaletteOptions & SocialPaletteOptions;
-    functions?: Functions;
   }
 
   interface PaletteOptions {
